Fix warn/report flow

+++ src/middlewares/delete-message.ts
@@ -1,6 +1,8 @@
 import { Middleware } from 'telegraf';
 import { Ctx } from '../types';
 
-export const deleteMessage: Middleware<Ctx> = async function(ctx) {
-  ctx.deleteMessage().catch();
+export const deleteMessage: Middleware<Ctx> = async function (ctx) {
+  if (ctx.dbChat?.delete_slash_commands) {
+    ctx.deleteMessage().catch();
+  }
 };
